+++ Bothtech/CustomUINavigationController/SquaresFlipNavigation/FlipSquaresNavigationController.m
@@ -344,8 +344,8 @@
     NSUInteger count = [array count];
     for (NSUInteger i = 0; i < count; ++i) {
         // Select a random element between i and end of array to swap with.
-        int nElements = count - i;
-        int n = (random() % nElements) + i;
+        NSInteger nElements = count - i;
+        NSInteger n = (random() % nElements) + i;
         [array exchangeObjectAtIndex:i withObjectAtIndex:n];
     }
